Coverage for interval confidence p

This function calculates the true coverage for any interval confidence for p.

ci_p_coverage(n, p, conf.level = 0.95, intervalType = "wald")

Arguments

n

number of trials.

p

true value for p.

conf.level

nominal confidence level for the returned confidence interval. By default is 0.95.

intervalType

type of confidence interval, possible choices are listed in ci_p.

Value

A dataframe with Method, n, p and true coverage.

Details

This function was inspired by the binomTestCoveragePlot() function from conf package and Park & Leemis (2019).

References

Park, H., & Leemis, L. M. (2019). Ensemble confidence intervals for binomial proportions. Statistics in Medicine, 38(18), 3460-3475.

See also

Author

David Esteban Cartagena Mejía, dcartagena@unal.edu.co

Examples

ci_p_coverage(n=10, p=0.45, intervalType="wald")
#>   intervalType  n    p  coverage
#> 1    ci_p_wald 10 0.45 0.8730485
ci_p_coverage(n=10, p=0.45, intervalType="wilson")
#>   intervalType  n    p  coverage
#> 1  ci_p_wilson 10 0.45 0.9493511
ci_p_coverage(n=10, p=c(0.10, 0.25, 0.40), intervalType="wilson")
#>   intervalType  n    p  coverage
#> 1  ci_p_wilson 10  0.1 0.9298092
#> 2  ci_p_wilson 10 0.25 0.9802723
#> 3  ci_p_wilson 10  0.4 0.9816588