Test for two mean vectors

This function implements the test for \(H0: \mu 1 = \mu 2\) versus \(H1: \mu1\) not = \(\mu2\). Muchachos, por favor revisar si los argumentos estan bien escritos.

two_mean_vector_test(xbar1, Sigma1, n1, xbar2, Sigma2, n2, delta0 = NULL,
  alpha = 0.05, method = "T2")

Arguments

xbar1
a vector with the sample mean from population 1.

Sigma1
a matrix with sample variances and covariances from population 1.

n1
sample size 1.

xbar2
a vector with the sample mean from population 2.

Sigma2
a matrix with sample variances and covariances from population 2.

n2
sample size 2.

delta0
que es esto

alpha
yo creo que este parametro lo debemos quita, que opinan?

method
tipo de prueba de hipotesis a usar. Las opciones disponibles son T2, bla bla. Por defecto es T2.

Value

por ahora retorna una lista de cosas, debemos hacer que la salida sea htest y que entregue lo que se necesita.

Examples

# Ejemplo 3.5.2 de Diaz y Morales (2005).
# Valentina y Jean Paul, debemos buscar mejor un ejemplo
# de un libro en ingles.
x1_barra <- c(15.97, 15.91, 27.19, 22.75)
x2_barra <- c(12.34, 13.91, 16.59, 21.94)
s1 <- matrix(c(5.192, 4.545, 6.522, 5.25, 4.545, 13.184, 6.76, 6.266, 6.522,
               6.76, 28.67, 14.468, 5.25, 6.266, 14.468, 16.645), ncol = 4)
s2 <- matrix(c(9.136, 7.549, 5.531, 4.151, 7.549, 18.6, 5.446, 5.446, 5.531,
               5.446, 13.55, 13.55, 4.151, 5.446, 13.55, 28), ncol = 4)

two_mean_vector_test(xbar1=x1_barra, Sigma1=s1, n1=32,
                     xbar2=x2_barra, Sigma2=s2, n2=32)

#> $sp
#>        [,1]   [,2]    [,3]    [,4]
#> [1,] 7.1640  6.047  6.0265  4.7005
#> [2,] 6.0470 15.892  6.1030  5.8560
#> [3,] 6.0265  6.103 21.1100 14.0090
#> [4,] 4.7005  5.856 14.0090 22.3225
#> 
#> $T2
#> [1] 136.6716
#> 
#> $p.value
#> [1] 2.563318e-14
#> 
#> $critic.value
#> [1] 10.62578
#> 

# Example 5.4.2 from Rencher & Christensen (2012) page 137.
x1_barra <- c(15.97, 15.91, 27.19, 22.75)
x2_barra <- c(12.34, 13.91, 16.66, 21.94)
s1 <- matrix(c(5.192, 4.545, 6.522, 5.25, 4.545, 13.18, 6.76, 6.266, 6.522,
               6.76, 28.67, 14.47, 5.25, 6.266, 14.47, 16.65), ncol = 4)
s2 <- matrix(c(9.136, 7.549, 4.864, 4.151, 7.549, 18.6, 10.22, 5.446, 4.864,
               10.22, 30.04, 13.49, 4.151, 5.446, 13.49, 28), ncol = 4)

two_mean_vector_test(xbar1=x1_barra, Sigma1=s1, n1=32,
                     xbar2=x2_barra, Sigma2=s2, n2=32, alpha = 0.01)

#> $sp
#>        [,1]   [,2]   [,3]    [,4]
#> [1,] 7.1640  6.047  5.693  4.7005
#> [2,] 6.0470 15.890  8.490  5.8560
#> [3,] 5.6930  8.490 29.355 13.9800
#> [4,] 4.7005  5.856 13.980 22.3250
#> 
#> $T2
#> [1] 97.67827
#> 
#> $p.value
#> [1] 1.444111e-11
#> 
#> $critic.value
#> [1] 15.36323
#> 

# Example 3.7 from Seber (1984) page 116.
# using the James first order test (1954).
n1 <- 16
xb1 <- c(9.82, 15.06)
s1 <- matrix(c(120, -16.3, -16.3, 17.8), ncol = 2)

n2 <- 11
xb2 <- c(13.05, 22.57)
s2 <- matrix(c(81.8, 32.1, 32.1, 53.8), ncol = 2)
two_mean_vector_test(xbar1 = xb1, Sigma1 = s1, n1 = n1,
                     xbar2 = xb2, Sigma2 = s2, n2 = n2,
                     method = 'james')

#> $T2
#> [1] 9.445541
#> 
#> $p.value
#> [1] 0.02046975
#>