Coverage for interval confidence p

This function calculates the true coverage for any interval confidence for p.

ci_p_coverage(n, p, conf.level = 0.95, intervalType = "wald")

Arguments

n

number of trials.

p

true value for p.

conf.level

nominal confidence level for the returned confidence interval. By default is 0.95.

intervalType

type of confidence interval, possible choices are listed in ci_p.

Value

A dataframe with Method, n, p and true coverage.

Details

This function was inspired by the binomTestCoveragePlot() function from conf package and Park & Leemis (2019).

References

Park, H., & Leemis, L. M. (2019). Ensemble confidence intervals for binomial proportions. Statistics in Medicine, 38(18), 3460-3475.

See also

Author

David Esteban Cartagena Mejía, dcartagena@unal.edu.co

Examples

ci_p_coverage(n=10, p=0.45, intervalType="wald")
#>      Method  n    p  coverage
#> 1 ci_p_wald 10 0.45 0.8730485
ci_p_coverage(n=10, p=0.45, intervalType="wilson")
#>        Method  n    p  coverage
#> 1 ci_p_wilson 10 0.45 0.9493511
ci_p_coverage(n=10, p=c(0.10, 0.25, 0.40), intervalType="wilson")
#>        Method  n    p  coverage
#> 1 ci_p_wilson 10 0.10 0.9298092
#> 2 ci_p_wilson 10 0.25 0.9802723
#> 3 ci_p_wilson 10 0.40 0.9816588
ci_p_coverage(n=10, intervalType="wilson")
#>          Method  n           p  coverage
#> 1   ci_p_wilson 10 0.000000000 1.0000000
#> 2   ci_p_wilson 10 0.005025126 0.9508700
#> 3   ci_p_wilson 10 0.010050251 0.9039231
#> 4   ci_p_wilson 10 0.015075377 0.8590728
#> 5   ci_p_wilson 10 0.020100503 0.9836681
#> 6   ci_p_wilson 10 0.025125628 0.9751572
#> 7   ci_p_wilson 10 0.030150754 0.9651738
#> 8   ci_p_wilson 10 0.035175879 0.9538533
#> 9   ci_p_wilson 10 0.040201005 0.9413234
#> 10  ci_p_wilson 10 0.045226131 0.9277045
#> 11  ci_p_wilson 10 0.050251256 0.9131105
#> 12  ci_p_wilson 10 0.055276382 0.8976482
#> 13  ci_p_wilson 10 0.060301508 0.9809078
#> 14  ci_p_wilson 10 0.065326633 0.9763748
#> 15  ci_p_wilson 10 0.070351759 0.9712831
#> 16  ci_p_wilson 10 0.075376884 0.9656274
#> 17  ci_p_wilson 10 0.080402010 0.9594061
#> 18  ci_p_wilson 10 0.085427136 0.9526208
#> 19  ci_p_wilson 10 0.090452261 0.9452762
#> 20  ci_p_wilson 10 0.095477387 0.9373798
#> 21  ci_p_wilson 10 0.100502513 0.9289413
#> 22  ci_p_wilson 10 0.105527638 0.9199728
#> 23  ci_p_wilson 10 0.110552764 0.9818939
#> 24  ci_p_wilson 10 0.115577889 0.9789234
#> 25  ci_p_wilson 10 0.120603015 0.9756525
#> 26  ci_p_wilson 10 0.125628141 0.9720710
#> 27  ci_p_wilson 10 0.130653266 0.9681697
#> 28  ci_p_wilson 10 0.135678392 0.9639406
#> 29  ci_p_wilson 10 0.140703518 0.9593765
#> 30  ci_p_wilson 10 0.145728643 0.9544714
#> 31  ci_p_wilson 10 0.150753769 0.9492203
#> 32  ci_p_wilson 10 0.155778894 0.9436194
#> 33  ci_p_wilson 10 0.160804020 0.9376659
#> 34  ci_p_wilson 10 0.165829146 0.9313580
#> 35  ci_p_wilson 10 0.170854271 0.9828394
#> 36  ci_p_wilson 10 0.175879397 0.9806313
#> 37  ci_p_wilson 10 0.180904523 0.9782303
#> 38  ci_p_wilson 10 0.185929648 0.9756281
#> 39  ci_p_wilson 10 0.190954774 0.9728170
#> 40  ci_p_wilson 10 0.195979899 0.9697894
#> 41  ci_p_wilson 10 0.201005025 0.9665380
#> 42  ci_p_wilson 10 0.206030151 0.9630559
#> 43  ci_p_wilson 10 0.211055276 0.9593366
#> 44  ci_p_wilson 10 0.216080402 0.9553738
#> 45  ci_p_wilson 10 0.221105528 0.9511620
#> 46  ci_p_wilson 10 0.226130653 0.9466958
#> 47  ci_p_wilson 10 0.231155779 0.9419703
#> 48  ci_p_wilson 10 0.236180905 0.9369813
#> 49  ci_p_wilson 10 0.241206030 0.9834809
#> 50  ci_p_wilson 10 0.246231156 0.9816995
#> 51  ci_p_wilson 10 0.251256281 0.9797787
#> 52  ci_p_wilson 10 0.256281407 0.9777121
#> 53  ci_p_wilson 10 0.261306533 0.9754934
#> 54  ci_p_wilson 10 0.266331658 0.9731163
#> 55  ci_p_wilson 10 0.271356784 0.9705748
#> 56  ci_p_wilson 10 0.276381910 0.9678625
#> 57  ci_p_wilson 10 0.281407035 0.9282597
#> 58  ci_p_wilson 10 0.286432161 0.9276761
#> 59  ci_p_wilson 10 0.291457286 0.9267512
#> 60  ci_p_wilson 10 0.296482412 0.9254879
#> 61  ci_p_wilson 10 0.301507538 0.9238886
#> 62  ci_p_wilson 10 0.306532663 0.9219555
#> 63  ci_p_wilson 10 0.311557789 0.9196910
#> 64  ci_p_wilson 10 0.316582915 0.9632168
#> 65  ci_p_wilson 10 0.321608040 0.9633911
#> 66  ci_p_wilson 10 0.326633166 0.9633591
#> 67  ci_p_wilson 10 0.331658291 0.9631216
#> 68  ci_p_wilson 10 0.336683417 0.9626789
#> 69  ci_p_wilson 10 0.341708543 0.9620312
#> 70  ci_p_wilson 10 0.346733668 0.9611780
#> 71  ci_p_wilson 10 0.351758794 0.9601188
#> 72  ci_p_wilson 10 0.356783920 0.9588526
#> 73  ci_p_wilson 10 0.361809045 0.9573783
#> 74  ci_p_wilson 10 0.366834171 0.9556946
#> 75  ci_p_wilson 10 0.371859296 0.9537996
#> 76  ci_p_wilson 10 0.376884422 0.9516917
#> 77  ci_p_wilson 10 0.381909548 0.9493688
#> 78  ci_p_wilson 10 0.386934673 0.9468287
#> 79  ci_p_wilson 10 0.391959799 0.9440692
#> 80  ci_p_wilson 10 0.396984925 0.9819747
#> 81  ci_p_wilson 10 0.402010050 0.9814254
#> 82  ci_p_wilson 10 0.407035176 0.9438728
#> 83  ci_p_wilson 10 0.412060302 0.9453904
#> 84  ci_p_wilson 10 0.417085427 0.9466723
#> 85  ci_p_wilson 10 0.422110553 0.9477206
#> 86  ci_p_wilson 10 0.427135678 0.9485368
#> 87  ci_p_wilson 10 0.432160804 0.9491223
#> 88  ci_p_wilson 10 0.437185930 0.9494782
#> 89  ci_p_wilson 10 0.442211055 0.9496052
#> 90  ci_p_wilson 10 0.447236181 0.9495040
#> 91  ci_p_wilson 10 0.452261307 0.9491746
#> 92  ci_p_wilson 10 0.457286432 0.9486173
#> 93  ci_p_wilson 10 0.462311558 0.9478315
#> 94  ci_p_wilson 10 0.467336683 0.9468170
#> 95  ci_p_wilson 10 0.472361809 0.9455728
#> 96  ci_p_wilson 10 0.477386935 0.9440980
#> 97  ci_p_wilson 10 0.482412060 0.9423914
#> 98  ci_p_wilson 10 0.487437186 0.9404515
#> 99  ci_p_wilson 10 0.492462312 0.9783757
#> 100 ci_p_wilson 10 0.497487437 0.9785001
#> 101 ci_p_wilson 10 0.502512563 0.9785001
#> 102 ci_p_wilson 10 0.507537688 0.9783757
#> 103 ci_p_wilson 10 0.512562814 0.9404515
#> 104 ci_p_wilson 10 0.517587940 0.9423914
#> 105 ci_p_wilson 10 0.522613065 0.9440980
#> 106 ci_p_wilson 10 0.527638191 0.9455728
#> 107 ci_p_wilson 10 0.532663317 0.9468170
#> 108 ci_p_wilson 10 0.537688442 0.9478315
#> 109 ci_p_wilson 10 0.542713568 0.9486173
#> 110 ci_p_wilson 10 0.547738693 0.9491746
#> 111 ci_p_wilson 10 0.552763819 0.9495040
#> 112 ci_p_wilson 10 0.557788945 0.9496052
#> 113 ci_p_wilson 10 0.562814070 0.9494782
#> 114 ci_p_wilson 10 0.567839196 0.9491223
#> 115 ci_p_wilson 10 0.572864322 0.9485368
#> 116 ci_p_wilson 10 0.577889447 0.9477206
#> 117 ci_p_wilson 10 0.582914573 0.9466723
#> 118 ci_p_wilson 10 0.587939698 0.9453904
#> 119 ci_p_wilson 10 0.592964824 0.9438728
#> 120 ci_p_wilson 10 0.597989950 0.9814254
#> 121 ci_p_wilson 10 0.603015075 0.9819747
#> 122 ci_p_wilson 10 0.608040201 0.9440692
#> 123 ci_p_wilson 10 0.613065327 0.9468287
#> 124 ci_p_wilson 10 0.618090452 0.9493688
#> 125 ci_p_wilson 10 0.623115578 0.9516917
#> 126 ci_p_wilson 10 0.628140704 0.9537996
#> 127 ci_p_wilson 10 0.633165829 0.9556946
#> 128 ci_p_wilson 10 0.638190955 0.9573783
#> 129 ci_p_wilson 10 0.643216080 0.9588526
#> 130 ci_p_wilson 10 0.648241206 0.9601188
#> 131 ci_p_wilson 10 0.653266332 0.9611780
#> 132 ci_p_wilson 10 0.658291457 0.9620312
#> 133 ci_p_wilson 10 0.663316583 0.9626789
#> 134 ci_p_wilson 10 0.668341709 0.9631216
#> 135 ci_p_wilson 10 0.673366834 0.9633591
#> 136 ci_p_wilson 10 0.678391960 0.9633911
#> 137 ci_p_wilson 10 0.683417085 0.9632168
#> 138 ci_p_wilson 10 0.688442211 0.9196910
#> 139 ci_p_wilson 10 0.693467337 0.9219555
#> 140 ci_p_wilson 10 0.698492462 0.9238886
#> 141 ci_p_wilson 10 0.703517588 0.9254879
#> 142 ci_p_wilson 10 0.708542714 0.9267512
#> 143 ci_p_wilson 10 0.713567839 0.9276761
#> 144 ci_p_wilson 10 0.718592965 0.9282597
#> 145 ci_p_wilson 10 0.723618090 0.9678625
#> 146 ci_p_wilson 10 0.728643216 0.9705748
#> 147 ci_p_wilson 10 0.733668342 0.9731163
#> 148 ci_p_wilson 10 0.738693467 0.9754934
#> 149 ci_p_wilson 10 0.743718593 0.9777121
#> 150 ci_p_wilson 10 0.748743719 0.9797787
#> 151 ci_p_wilson 10 0.753768844 0.9816995
#> 152 ci_p_wilson 10 0.758793970 0.9834809
#> 153 ci_p_wilson 10 0.763819095 0.9369813
#> 154 ci_p_wilson 10 0.768844221 0.9419703
#> 155 ci_p_wilson 10 0.773869347 0.9466958
#> 156 ci_p_wilson 10 0.778894472 0.9511620
#> 157 ci_p_wilson 10 0.783919598 0.9553738
#> 158 ci_p_wilson 10 0.788944724 0.9593366
#> 159 ci_p_wilson 10 0.793969849 0.9630559
#> 160 ci_p_wilson 10 0.798994975 0.9665380
#> 161 ci_p_wilson 10 0.804020101 0.9697894
#> 162 ci_p_wilson 10 0.809045226 0.9728170
#> 163 ci_p_wilson 10 0.814070352 0.9756281
#> 164 ci_p_wilson 10 0.819095477 0.9782303
#> 165 ci_p_wilson 10 0.824120603 0.9806313
#> 166 ci_p_wilson 10 0.829145729 0.9828394
#> 167 ci_p_wilson 10 0.834170854 0.9313580
#> 168 ci_p_wilson 10 0.839195980 0.9376659
#> 169 ci_p_wilson 10 0.844221106 0.9436194
#> 170 ci_p_wilson 10 0.849246231 0.9492203
#> 171 ci_p_wilson 10 0.854271357 0.9544714
#> 172 ci_p_wilson 10 0.859296482 0.9593765
#> 173 ci_p_wilson 10 0.864321608 0.9639406
#> 174 ci_p_wilson 10 0.869346734 0.9681697
#> 175 ci_p_wilson 10 0.874371859 0.9720710
#> 176 ci_p_wilson 10 0.879396985 0.9756525
#> 177 ci_p_wilson 10 0.884422111 0.9789234
#> 178 ci_p_wilson 10 0.889447236 0.9818939
#> 179 ci_p_wilson 10 0.894472362 0.9199728
#> 180 ci_p_wilson 10 0.899497487 0.9289413
#> 181 ci_p_wilson 10 0.904522613 0.9373798
#> 182 ci_p_wilson 10 0.909547739 0.9452762
#> 183 ci_p_wilson 10 0.914572864 0.9526208
#> 184 ci_p_wilson 10 0.919597990 0.9594061
#> 185 ci_p_wilson 10 0.924623116 0.9656274
#> 186 ci_p_wilson 10 0.929648241 0.9712831
#> 187 ci_p_wilson 10 0.934673367 0.9763748
#> 188 ci_p_wilson 10 0.939698492 0.9809078
#> 189 ci_p_wilson 10 0.944723618 0.8976482
#> 190 ci_p_wilson 10 0.949748744 0.9131105
#> 191 ci_p_wilson 10 0.954773869 0.9277045
#> 192 ci_p_wilson 10 0.959798995 0.9413234
#> 193 ci_p_wilson 10 0.964824121 0.9538533
#> 194 ci_p_wilson 10 0.969849246 0.9651738
#> 195 ci_p_wilson 10 0.974874372 0.9751572
#> 196 ci_p_wilson 10 0.979899497 0.9836681
#> 197 ci_p_wilson 10 0.984924623 0.8590728
#> 198 ci_p_wilson 10 0.989949749 0.9039231
#> 199 ci_p_wilson 10 0.994974874 0.9508700
#> 200 ci_p_wilson 10 1.000000000 0.0000000