Tests for Equality of Two Normal Mean Vectors

Implements the test for \(H_0: \mu_1 = \mu_2\) versus \(H_1: \mu_1\) not = \(\mu_2\) when both random samples are from two p-variate normal populations \(Np(\mu_1, \Sigma_1)\) and \(Np(\mu_2, \Sigma_2)\). By default, this function performs the Hotelling test for two normal mean vectors assumming equality in the covariance matrices. Also testing the multivariate Behrens-Fisher problem, when the assumption of equal covariance matrices is violated, including James, Yao, Johansen, ... approaches.

two_mean_vector_test(xbar1, Sigma1, n1, xbar2, Sigma2, n2, delta0 = NULL,
  alpha = 0.05, method = "T2")

Arguments

xbar1
a vector with the sample mean from population 1.

Sigma1
a matrix with sample variances and covariances from population 1.

n1
sample size 1.

xbar2
a vector with the sample mean from population 2.

Sigma2
a matrix with sample variances and covariances from population 2.

n2
sample size 2.

delta0
a number indicating the true value of the difference in means.

alpha
the significance level, by default its value is 0.05.

method
a character string specifying the method, it must be one of "T2" (default), "james" (James first order test), "yao" (Yao test), "johansen" (Johansen test), ...

Value

A list with class "htest" containing the following components:

statistic

the value of the statistic.

parameter

the degrees of freedom for the test.

p.value

the p-value for the test.

estimate

the estimated mean vectors.

method

a character string indicating the type of test performed.

Examples

# Example 5.4.2 from Rencher & Christensen (2012) page 137.
n1 <- 32
xbar1 <- c(15.97, 15.91, 27.19, 22.75)
s1 <- matrix(c(5.192, 4.545, 6.522, 5.25, 4.545, 13.18, 6.76, 6.266, 6.522,
               6.76, 28.67, 14.47, 5.25, 6.266, 14.47, 16.65), ncol = 4)

n2 <- 32
xbar2 <- c(12.34, 13.91, 16.66, 21.94)
s2 <- matrix(c(9.136, 7.549, 4.864, 4.151, 7.549, 18.6, 10.22, 5.446, 4.864,
               10.22, 30.04, 13.49, 4.151, 5.446, 13.49, 28), ncol = 4)

two_mean_vector_test(xbar1 = xbar1, Sigma1 = s1, n1 = n1,
                     xbar2 = xbar2, Sigma2 = s2, n2 = n2, alpha = 0.01)

#> 
#> 	T2 test for two mean vectors
#> 
#> data:  this test uses summarized data
#> T2 = 97.678, F = 23.238, df1 = 4, df2 = 59, p-value = 1.444e-11
#> alternative hypothesis: mu1 is not equal to mu2 
#> 
#> sample estimates:
#>        Sample 1 Sample 2
#> xbar_1    15.97    12.34
#> xbar_2    15.91    13.91
#> xbar_3    27.19    16.66
#> xbar_4    22.75    21.94
#> 

# Example 3.7 from Seber (1984) page 116.
# using the James first order test (1954).
n1 <- 16
xbar1 <- c(9.82, 15.06)
s1 <- matrix(c(120, -16.3, -16.3, 17.8), ncol = 2)

n2 <- 11
xbar2 <- c(13.05, 22.57)
s2 <- matrix(c(81.8, 32.1, 32.1, 53.8), ncol = 2)

two_mean_vector_test(xbar1 = xbar1, Sigma1 = s1, n1 = n1,
                     xbar2 = xbar2, Sigma2 = s2, n2 = n2,
                     method = 'james')

#> 
#> 	James test for two mean vectors
#> 
#> data:  this test uses summarized data
#> T2 = 9.4455, X-squared = 7.7776, df = 2, p-value = 0.02047
#> alternative hypothesis: mu1 is not equal to mu2 
#> 
#> sample estimates:
#>        Sample 1 Sample 2
#> xbar_1     9.82    13.05
#> xbar_2    15.06    22.57
#> 

# Example 4.1 from Nel and Van de Merwe (1986) page 3729
# Test H0: mu1 = mu2 versus H1: mu1 != mu2
n1 <- 45
xbar1 <- c(204.4, 556.6)
s1 <- matrix(c(13825.3, 23823.4, 23823.4, 73107.4), ncol=2)

n2 <- 55
xbar2 <- c(130.0, 355.0)
s2 <- matrix(c(8632.0, 19616.7, 19616.7, 55964.5), ncol=2)

two_mean_vector_test(xbar1 = xbar1, Sigma1 = s1, n1 = n1,
                     xbar2 = xbar2, Sigma2 = s2, n2 = n2,
                     method = 'mvn')

#> 
#> 	Modified Nel and Van der Merwe test for two mean vectors
#> 
#> data:  this test uses summarized data
#> T2 = 15.6585, F = 7.7284, df1 = 2.000, df2 = 76.592, p-value =
#> 0.0008763
#> alternative hypothesis: mu1 is not equal to mu2 
#> 
#> sample estimates:
#>        Sample 1 Sample 2
#> xbar_1    204.4      130
#> xbar_2    556.6      355
#>